Add files via upload
Team Data per year from 2002-2017. One weird thing is that some team names have "NCAA" next to them if they made the tournament. For instance, the name of Duke is always "Duke NCAA" instead of just Duke, so that will need to be accounted for. The most important stats are probably win-loss (W-L%) and points, both Tm. (the school) and Opp. (how much opponents scored).